requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
cryptography==38.0.3
cssbeautifier==1.14.7
Django==4.1.2
django-countries==7.4.2
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.14.0
djlint==1.19.7
EditorConfig==0.12.3
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.2
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
jsbeautifier==1.14.7
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.3.0
platformdirs==2.5.4
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
